pyproject.toml
[project]
name = "memsrv"
version = "0.1.0"
description = "A simple, self-hosted memory service boilerplate for LLMs and agent frameworks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "chromadb>=1.1.0",
    "fastapi>=0.116.2",
    "google-genai>=1.38.0",
    "openinference-semantic-conventions>=0.1.23",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
examples-adk = [
    "google-adk>=1.14.1",
    "streamlit>=1.49.1",
]
examples-langchain = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
]
